// src/app/interfaces/pedido.ts
export interface Pedido {
    NumeroOrden: any;
    idAlmacen: string; 
    NitProveedor: string;
    FechaPedido: any;
    ID_Registro: string;
    NumeroOrden_Aux: string;
    IdPlu: any;
    CodigoLargo:string;
    CodigoEAN :string ;
    Cantidad: number;
    PrecioFacturacion:string; 
    PrecioPublico:string;
}
  